feature(nlp): moved logic out of controller into the service
